Drag and drop to sidebar
Also, when drag/dropping tasks, they will clone if dragged from a daily
plan, and will not when dragged from a tickler or other list. It's a first
step towards having pinned items.
Note: this step currently depends on changes to vue-draggable. Will be
working to get those merged, or fork my own copy of it.